refactor: Extract analysis, capture, and error views (Issue #3 Phase 5)
Created three new specialized modules:
- analysis.rs (529 lines) - Analysis results (overview, devices, security, timing)
- capture.rs (307 lines) - Capture details and list table
- errors.rs (290 lines) - Error messages and comparison results

Result: core.rs reduced from 1,626 → 564 lines (1,062 lines extracted)
Cumulative: 5,020 → 564 lines (4,456 lines extracted, 88.8% reduction)

All functions use pub(crate) visibility for internal module use.
Module coordination through ui/mod.rs with proper imports in core.rs.
